+++ src/app/plan/page.tsx
@@ -1,5 +1,22 @@
-import TravelPage from "./TravelPage";
+import BackButton from "@/components/feature/backButton";
+import PlanRegion from "@/components/form/plan/planRegion";
 
+// 여행계획
 export default function PlanPage() {
-  return <TravelPage />;
+
+
+  return (
+    <div>
+      <div className="w-full relative py-5 px-4">
+        <BackButton />
+        <p className="text-center">여행일정만들기</p>
+      </div>
+      <div className="relative w-full px-4 pb-25">
+        <h2 className="text-28 text-travel-primary200 font-semibold">
+          어디로 여행을 떠나시나요?
+        </h2>
+        <PlanRegion />
+      </div>
+    </div>
+  );
 }
